move widget ID utils
